requirements.txt
torch==1.10.0+cu113
torchvision==0.11.1+cu113
torchaudio===0.10.0+cu113
Pillow===8.4.0
numpy==1.21.4
